+++ test/unit/index.js
@@ -15,7 +15,7 @@ describe('use component', () => {
 });
 Vue.use(kingdot);
 
-const testsContext = require.context('./specs', true, /\.spec$/);
+const testsContext = require.context('./spec', true, /\button.spec$/);
 
 
 testsContext.keys().forEach(testsContext);