+++ CHANGELOG.md
@@ -2,6 +2,14 @@
     
 ## v1.0.5 
 
+*2021-04-02*
+
+### New features
+### Bug fixes
+### Optimization
+* Tabs 
+    * 测试用例优化
+
 *2021-04-01*
 
 ### New features