+++ CHANGELOG.md
@@ -10,7 +10,10 @@
 * transfer
     * 新增全局禁用功能
 ### Bug fixes
-  
+* Dialog
+    * 测试用例修改
+* Tabs 
+    * 测试用例修改
 ### Optimization
 * Dialog
     * 滚动条增加宽度样式
@@ -38,5 +41,3 @@
 
 * Docs
     * Button 修改loading示例
-* Dialog
-    * 滚动条增加宽度样式