+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Bug fixes
 * Select
     * 修改延迟赋值value不显示label   
+    * 添加单测用例
 ## v1.0.5 
 
 *2021-04-01*