+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * Select
     * 修改延迟赋值value不显示label   
     * 添加单测用例
+* Button
+    * 默认原生type修改成button
 ## v1.0.5 
 
 *2021-04-01*